refact: broke SavedRecipeCard component into smaller subcomponents for improved readability and maintainability
